refactor: created a file for each endpoint and added custom error handling
