--- bricklink_py/config.py
@@ -1,8 +1,0 @@
-# Bricklink tokens
-CONSUMER_KEY = None
-CONSUMER_SECRET = None
-TOKEN = None
-TOKEN_SECRET = None
-
-# Api URLs
-API_BASE_URL = 'https://api.bricklink.com/api/store/v1/'
